debian/rules: run tests only once, with verbose output
To make build logs easier to read, debian/rules runs tests once with
terse output and then runs tests again to reproduce any failures with
more verbose output.  The idea is that in the normal case the verbose
output from tests is added noise that is distracting when reading
build logs.

However, that approach has a few downsides for automated builds:

 * since it runs tests twice on failure, it can mask flaky tests
 * if a test hangs and the build times out, there is no log describing
   what happened except for the name of the test
 * it is hard to compare logs from successful builds to unsuccessful
   ones to track down a failure

An analogous situation exists in the build rules: the Debian build
turns on V=1 to ensure the build log continues all commands used to
build the source code, not just failed commands.

Do the analogous thing for tests: use verbose output in the first
place and these complexities and pitfalls go away.

Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>