Refer to the FRAME .tf through a variable tfpath.
This is the top frame:

   ( =====================
tf:(  Top half (history list)
   ( =====================
              |
      diff    | file-list
    =====================

Introduce a global variable tfpath, and operate with that instead of ".tf"

It's not passed to functions as parameter yet.

$tf (hierarchy address $tfpath)

We gain flexibility of the widget hierarchy in code.  -- do we?

Careful in callbacks: tf variable needs to be (lexically re-) defined